Create README.md
The Hospital Management System is a application developed using Java and MySQL, designed to 
efficiently manage various aspects of a healthcare facility. Java is utilized for creating a user-friendly 
command line interface, and the backend logic, while MySQL serves as the relational database for storing 
and retrieving patient information, appointments, doctor details.